refactor: useResponsiveGrid 분리
